Split eval.c into three separate files
This commit refactors the expression evaluator (eval.c) by splitting it into three
separate files:

- eval-parser.c: Parser implementation (recursive descent parser)
- eval-execute.c: Execution and evaluation logic
- eval-utils.c: Common utilities and memory management

The original eval.c file is now a stub for backward compatibility.
This refactoring doesn't change any functionality, but improves code
organization and will facilitate future improvements to the expression
parser.
